+++ packages/_private/react-demo/src/main.tsx
@@ -2,6 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import '@synergy-design-system/tokens/themes/dark.css';
 import '@synergy-design-system/tokens/themes/light.css';
+import '@synergy-design-system/components/themes/utility.css';
 import './app.css';
 import { App } from './App';